+++ src/components/componentUserStatus/index.ts
@@ -1,5 +1,6 @@
 import { register } from '../../shared/functions';
 import ComponentUserStatus from './componentUserStatus';
 import '../componentSlider/index';
+import '../componentEmoji/index';
 
 register('component-user-status', ComponentUserStatus);