+++ src/components/componentEmoji/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-emoji
 
+## 0.0.3
+
+### Patch Changes
+
+- Updating ui for icons, home page, search input and cards in user settings page
+
 ## 0.0.2
 
 ### Patch Changes