+++ src/components/componentSearchInput/CHANGELOG.md
@@ -1,5 +1,11 @@
 # search-input
 
+## 0.0.2
+
+### Patch Changes
+
+- Updating ui for icons, home page, search input and cards in user settings page
+
 ## 0.0.1
 
 ### Patch Changes