+++ src/components/pageHome/CHANGELOG.md
@@ -1,5 +1,11 @@
 # home-page
 
+## 0.0.15
+
+### Patch Changes
+
+- Updating ui for icons, home page, search input and cards in user settings page
+
 ## 0.0.14
 
 ### Patch Changes