+++ src/components/pageSearch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # search-page
 
+## 0.0.7
+
+### Patch Changes
+
+- Updating ui for icons, home page, search input and cards in user settings page
+
 ## 0.0.6
 
 ### Patch Changes